pgs/chatbot.py

#!/usr/bin/env python3

import streamlit as st
import google.generativeai as genai
import os

from dotenv import load_dotenv

load_dotenv()

genai.configure(api_key = os.getenv("GOOGLE_API_KEY"))


def get_gemini_response(prompt):

    model = genai.GenerativeModel("gemini-1.5-flash", 

        system_instruction = """
        
            You are MediaBot a professional Kenyan journalist working for Nation Media Group. Your responses must always be grounded in Kenyan local news, including:

            - Breaking news

            - Articles

            - In-depth stories

            - Historical events

            - Political developments

            - Cultural matters

            - Economy and society

            You must only respond within the context of Kenya and Kenyan affairs. Do not answer questions outside this scope.

            When providing information or answering questions, always encourage users to refer to official and updated news from https://nation.africa/kenya for the most precise and verified reports.

            Maintain a professional, factual, and journalistic tone at all times.
            """

            )


    response = model.generate_content(
        prompt,
        generation_config = genai.GenerationConfig(
        max_output_tokens=1000,
        temperature=0.1, 
      )
    
    )


    
    return response.text




# Initialize session state for chat history
if "messages" not in st.session_state:
    st.session_state.messages = [{"role": "assistant", "content": "How may I help you?"}]

# Display chat history
for message in st.session_state.messages:

    with st.chat_message(message["role"]):
        st.markdown(message["content"])



if prompt := st.chat_input("How may I help?"):
    # Append user message
    st.session_state.messages.append({"role": "user", "content": prompt})
    with st.chat_message("user"):
        st.markdown(prompt)

    # Generate AI response
    chat_output = get_gemini_response(prompt)
    
    # Append AI response
    with st.chat_message("assistant"):
        st.markdown(chat_output)

    st.session_state.messages.append({"role": "assistant", "content": chat_output})



