requirements.txt
streamlit
python-dotenv
streamlit_option_menu 
streamlit_chat
africastalking
openai
requests
google-generativeai
langchain
langchain-community
langchain-core
langchain-google-genai
google-generativeai
bcrypt
captcha
python-dotenv
pyttsx3
st-social-media-links
tqdm
streamlit-tags
chromadb
faiss-cpu
pypdf
youtube-transcript-api
streamlit-audiorec
streamlit-scrollable-textbox

